fix(src): Change structure app and creating docker-compose and dockerfile with src and postgres
